+++ Imagine.Scenes/QuadraticSolver.cs
@@ -2,21 +2,21 @@ namespace Imagine.Scenes;
 
 public static class QuadraticSolver
 {
-	public static List<double> Solve(double a, double b, double c)
+	public static List<float> Solve(float a, float b, float c)
 	{
-		var d = (b * b) - (4D * a * c);
+		var d = (b * b) - (4F * a * c);
 
-		if (d < 0D)
+		if (d < 0F)
 		{
 			return [];
 		}
 
-		var sqrtD = double.Sqrt(d);
+		var sqrtD = float.Sqrt(d);
 
 		return
 		[
-			(-b - sqrtD) / (2D * a),
-			(-b + sqrtD) / (2D * a),
+			(-b - sqrtD) / (2F * a),
+			(-b + sqrtD) / (2F * a),
 		];
 	}
 }
